Refactored NamedThreadFactory to standalone class
