--- engine/src/main/java/org/antublue/test/engine/internal/util/ThrowableUtils.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2022-2023 Douglas Hoard
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.antublue.test.engine.internal.util;
-
-/**
- * Class to implement Throwable utilities
- */
-public final class ThrowableUtils {
-
-    /**
-     * Constructor
-     */
-    private ThrowableUtils() {
-        // DO NOTHING
-    }
-
-    /**
-     * Method to throw a checked (RuntimeException) if a Throwable it checked
-     *
-     * @param t t
-     */
-    public static void throwUnchecked(Throwable t) {
-        if (t instanceof RuntimeException) {
-            throw (RuntimeException) t;
-        } else {
-            throw new RuntimeException(t);
-        }
-    }
-}
